requirements.txt
altgraph==0.17
anyascii==0.1.7
astroid==2.5.2
attrs==21.2.0
Automat==20.2.0
baidu-aip==2.2.18.0
beautifulsoup4==4.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
django==2.2.14
django-filter==2.4.0
django-modelcluster==5.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-taggit==1.3.0
django-tinymce==3.0.2
django-treebeard==4.4
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
future==0.18.2
h2==3.2.0
hpack==3.0.0
html5lib==1.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.0.1
incremental==21.3.0
isort==5.8.0
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
l18n==2020.6.1
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.3
openpyxl==3.0.6
parsel==1.6.0
pefile==2021.5.13
Pillow==7.2.0
prettytable==2.0.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyecharts==1.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylint==2.7.4
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pytz==2020.1
pywin32-ctypes==0.2.0
queuelib==1.6.1
requests==2.25.1
Scrapy==2.5.0
service-identity==21.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
tablib==3.0.0
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typed-ast==1.4.2
typing-extensions==3.10.0.0
Unidecode==1.1.2
urllib3==1.26.2
w3lib==1.22.0
wagtail==2.11.3
wcwidth==0.2.5
webencodings==0.5.1
wfastcgi==3.0.0
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==3.4.1
zope.interface==5.4.0
